-change rtkplot recognize *.rnx for obs and nav file type. -add C40 and C41 in codeopt.c and requency. -Match Beidou EPH IOD with SSR iodcrc.

+++ app/rtknavi_qt/mondlg.cpp
@@ -371,7 +371,7 @@ void MonitorDialog::ShowRtk(void)
 	rtksvrlock(&rtksvr); // lock
 	
 	rtk=rtksvr.rtk;
-    thread=(unsigned long)rtksvr.thread;
+    thread=(uintptr_t)rtksvr.thread;
 	cycle=rtksvr.cycle;
 	state=rtksvr.state;
 	rtkstat=rtksvr.rtk.sol.stat;
